Add E2E tests (#5)
# Problem

Wally doesn't provide a great experience when iterating on a package
from a consumer (flipbook). I'm finding myself having to go
back-and-forth between flipbook and Storyteller to get everything lined
up. Instead, I want to introduce some end-to-end tests to make it easier
to validate that what flipbook expects is what Storyteller expects

# Solution

This PR got away from me a bit. The bulk of the work is introducing the
`e2e` folder with Stories to test behavior. I've also made some API
adjustments to account for bugs I hit along the way.

With this, all stories except Fusion and Hoarcekat are passing. Since
this PR is so large already, I'm going to merge as-is and work through
the remaining issues later on