Accommodate analysis errors in flipbook's recent PR (#24)
# Problem

Trying to update flipbook's docs and part of that involves analyzing
code samples. But that's currently failing because Storyteller's Story
and Storybook types are too strict

https://github.com/flipbook-labs/flipbook/pull/276

# Solution

Created new `LoadedStory` and`LoadedStorybook` types which have all the
implicitly defined properties that a user doesn't need to worry about.
`Story` and `Storybook` are now slimmed down