Add tests for React hooks (#31)
# Problem

Our React hooks aren't tested and there are some bugs I've been
encountering in Flipbook

# Solution

Test React hooks. 

I've also tacked on a change to `useStorybooks` to have it return both
available and unavailable storybooks. This will make it possible for
Flipbook to show which storybooks were not able to be loaded, along with
error info. No more cryptic warnings in the output

Resolves #30