Storybook groups (#50)
# Problem

Roblox's internal story format has a `group` property on storybooks to
provide a basic way to group storybooks together

# Solution

In Developer Storybook, the storybooks that are grouped together are
placed under a folder in the sidebar. In our case, we're just appending
the group to the name. Flipbook sorts alphabetically so this is good
enough

Split from #29