Discovery of orphans with better DataModel awareness (#57)
# Problem

I'm trying to fix the following issues in Flipbook with this change:

1. https://github.com/flipbook-labs/flipbook/issues/434
  
Orphaned stories will not show up unless there's at least one Storybook
in the experience. This is because `useStorybooks` needs to run first,
but does not surface when it's still processing descendants. This means
that we can not make a distinction between "instances have been
traversed and there are 0 available/unavailable storybooks" and
"storybooks are still being discovered, so we have no results yet."

2. https://github.com/flipbook-labs/flipbook/issues/278

There are general issues around rerenders not triggering for common
cases like adding/removing/renaming Story and Storybook modules. This
leads to cases like: Insert a ModuleScript, rename to `Foo.story`,
update Source to be a properly formatted story, and observe that nothing
changed in the Flipbook UI.

# Solution

For the first issue, I've updated `useStorybooks` to include an
`isLoading` value so we can wait for the list of available storybooks to
resolve before checking which story modules are orphans. I've also
introduced the `useOrphanedStoryModules` hook to be used in conjunction
with `useStorybooks`.

For the second issue, I've revised a lot of the logic around instance
discovery and built it on top of QueryDescendants. Right now the
selector pattern can't handle name pattern matching (like matching for
`.story` or `.storybook`), but it can at least be used to grab just the
slice of descendants that are ModuleScripts. No more need for a gigantic
`game:GetDescendants()` allocation! I also added logic (and tests) to
make sure `useOrphanedStoryModules` and `useStorybooks` will update when
any ModuleScript is added/removed/renamed, which should keep the
Flipbook UI more in sync with the DataModel