Instance collector (#59)
# Problem

Storyteller does not do a great job keeping track of Story and
Storybooks as they change throughout a Studio session.

Here's a selection of issues this PR aims to resolve:
1. https://github.com/flipbook-labs/flipbook/issues/434
2. https://github.com/flipbook-labs/flipbook/issues/278
3. https://github.com/flipbook-labs/flipbook/issues/108
4. https://github.com/flipbook-labs/storyteller/issues/55

These issues primarily stem from Storyteller not triggering re-renders
when the DataModel changes.

# Solution

So, admittedly, I may have gotten a little carried away. 

This is a huge PR but the main things to focus on are:

1. StorytellerStore has been built from the ground up to support all of
the issues mentioned above, along with an extensive suite of unit tests
to support it.
2. The existing React hooks have had their internals hoisted up into
StorytellerStore so that all we really wind up with is some calls to
`useSignalState` and filtering based on an optional parent

I've done a lot of testing locally (we also have a storybook now to view
how the hooks behave) and from what I'm able to see this change should
significantly reduce the number of re-renders we're seeing in Flipbook
on v1.2.0.

It should also significantly improve behavior around detecting
Storybooks and Story modules as they go through their lives in the
DataModel. We're now doing a lot better tracking of a ModuleScript's
ancestry, name, and source changes. Cases like inserting a ModuleScript,
filling out the Source, and renaming it to `Foo.storybook` now results
in this Storybook being immediately picked up, rather than having to
reload the entire Flipbook UI to get it to recompute.

I'm inclined to move forward with this PR, cut a 1.3.0 release, and then
fix forward for any issues that crop up.