Performance improvements for collections (#67)
# Problem

Performance is tanking hard from InstanceQueryStore's DescendantAdded
connection. It was trivial enough to revise this, but I'm also unhappy
with how InstanceQueryStore and StorytellerStore work together.

So part of the problem I want to address is just making a really solid
querying system that can do everything InstanceQueryStore does, along
with taking over the responsibilities of all of the `onInstanceVerb` and
`onModuleScriptVerb` from StorytellerStore


# Solution

I've introduced a new `query` function that can handle the same cases as
InstanceQueryStore and has events that can be listened to for Instances
being added/changed/removed.

A query is created for Storybook modules and Story modules, and all the
hairy logic in StorytellerStore has been offloaded to the query events