Update createStory to take an arbitrary object (#70)
# Problem

In #49 I found createStory to be too limiting since it only accepts
positional args for the story element and the packages used for
rendering. I also need access to `controls`

# Solution

I went ahead and make createStory much more flexible by taking in an
object that gets merged with the defaults. This should make story
creation in tests more flexible