Give all React code a single home (#71)
# Problem

The `hooks` folder currently holds all of our React hooks, and with #49
we'll also be introducing a React component to the codebase. All of
these files should live together, but putting a component under hooks
doesn't make sense

# Solution

I renamed `hooks` to `react` to better encapsulate all of our
React-related files. I also want to keep everything grouped together
like this so it becomes trivial to split out to a `StorytellerReact`
package in the future